[quant] EqualizationQConfig to store input activation/weights
Summary: Created an EqualizationQConfig specifically for equalization.
This is used to store the input activation observer and weight observer.
Due to a design change, we plan to separately insert the
EqualizationObservers separately from the QuantizationObservers, we
will no longer need the output observer field in the
_InputEqualizationObserver.

Test Plan: compiles

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 73397e4bc435dae13629c630ef9572924a3cc32b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59739