[wip] quantization: store safe_on_fbgemm flag on quantized conv
Summary:

This is a start of fixing the problems surfaced in #46749.
This particular PR only fixes a small part of this:
1. if a conv module is unsafe to run in fbgemm, we now persist this
information.
2. if we are in an fbgemm kernel and we detect that the current conv
packed params are tagged as unsafe, we throw an error.

For now, this PR is a WIP to get some early feedback if this is the
right direction, since iteration cost on this is high. In particular,
missing things here are:
* better unit testing
* serialization, verifying that this is BC
* getting all the conv callsites (currently just module + conv2d is handled)

Note: there were some potential improvements discussed on dynamically
dispatching to qnnpack if it is available and the flag is set.  This PR
does not attempt to solve this issue - it can be solved by future PRs.

Test Plan:

```
python test/test_quantization.py TestQuantizedOps.test_conv_reduce_range
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 2b949d01af875da49c5646cbf766b859ae9b7189
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59984